requirements.txt
discord.py 
json
aiohttp
asyncio